+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+* Support for automatic Spring AI configuration mapping from Heroku Managed Inference and Agents (MIA) environment variables. ([#000](https://github.com/heroku/heroku-buildpack-jvm-common/pull/000))
+
 
 ## [v172] - 2025-09-17